Keep sbt-metals single file
